requirements.txt
black==23.3.0
dask==2023.4.0
datumaro==0.3
fastparquet==2023.2.0
fire>=0.4.0
imagehash==4.3.1
numpy~=1.23.5
opencv-python>=4.5.2.54
pandas~=1.5.3
parameterized==0.8.1
Pillow>=9.1.1
polars==0.16.13
pre-commit>=3.1.1
psutil~=5.9.4
pytest-env==0.8.1
PyYAML>=6.0
requests~=2.28.2
setuptools~=67.6.1
tqdm==4.64.1
uuid==1.30
vaex==4.16.0
